SKE controller refactor
